backend/app/requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
google-auth
requests
google-genai
fastapi[all]
bcrypt
python-jose[cryptography]
python-dotenv
